rm typescript + cleanup
