+++ ionic.config.json
@@ -1,5 +1,5 @@
 {
-  "name": "myApp",
+  "name": "maraudes",
   "integrations": {},
   "type": "react"
 }
