+++ package.json
@@ -5,13 +5,6 @@
   "dependencies": {
     "@ionic/react": "^4.11.0",
     "@ionic/react-router": "^4.11.0",
-    "@types/jest": "^24.0.18",
-    "@types/node": "^12.7.12",
-    "@types/react": "^16.9.5",
-    "@types/react-dom": "^16.9.1",
-    "@types/react-router": "^5.1.1",
-    "@types/react-router-dom": "^5.1.0",
-    "@types/uuid": "^3.4.6",
     "date-fns": "^2.7.0",
     "graphql": "^14.5.8",
     "ionicons": "^4.6.3",
@@ -20,7 +13,6 @@
     "react-router": "^5.1.0",
     "react-router-dom": "^5.1.0",
     "react-scripts": "3.2.0",
-    "typescript": "3.6.3",
     "urql": "^1.6.3",
     "uuid": "^3.3.3"
   },
