--- src/comments.json
@@ -1,26 +1,0 @@
-[
-  {
-    "id": "1",
-    "from": "Tony",
-    "message": "Je ne l'ai pas vu aujourd'hui !",
-    "creationDate": "2019-10-22T11:54:44.461Z"
-  },
-  {
-    "id": "2",
-    "from": "Laurie",
-    "message": "Noté ! On essaiera de le rencontrer jeudi",
-    "creationDate": "2019-10-22T11:54:44.461Z"
-  },
-  {
-    "id": "3",
-    "from": "Tony",
-    "message": "Croisé en maraude aujourd'hui",
-    "creationDate": "2019-10-22T11:54:44.461Z"
-  },
-  {
-    "id": "4",
-    "from": "Laurie",
-    "message": "Parfait, merci beaucoup !",
-    "creationDate": "2019-10-22T11:54:44.461Z"
-  }
-]
