--- src/pages/Home.css
@@ -1,11 +1,0 @@
-.welcome-card img {
-  overflow: hidden;
-}
-
-.home-image {
-  background-image: url(/assets/glenn-carstens-peters-rTO4hjAgbDU-unsplash.jpg);
-  background-position: center center;
-  background-size: cover;
-  background-repeat: no-repeat;
-  height: 35vh;
-}
