--- src/persons.json
@@ -1,26 +1,0 @@
-[
-  {
-    "id": "1",
-    "title": "Olivier",
-    "creationDate": "2019-10-22T11:54:44.461Z",
-    "creator": "Laurie"
-  },
-  {
-    "id": "2",
-    "title": "Manon",
-    "creationDate": "2019-04-04T14:20:44.461Z",
-    "creator": "Laurie"
-  },
-  {
-    "id": "3",
-    "title": "Mimi",
-    "creationDate": "2018-12-03T21:43:44.461Z",
-    "creator": "Laurie"
-  },
-  {
-    "id": "4",
-    "title": "M des Batignolles",
-    "creationDate": "2019-01-12T12:01:44.461Z",
-    "creator": "Laurie"
-  }
-]
