--- src/todo.json
@@ -1,44 +1,0 @@
-[
-    {
-        "id": "1",
-        "person": "Olivier",
-        "category": "sante",
-        "title": "Acte IDE plaie genou",
-        "dueDate": "2019-12-08T11:54:44.461Z",
-        "creationDate": "2019-10-22T11:54:44.461Z",
-        "description": "Some description of the task and more details if needed like location, links or whats needed to compelte the task",
-        "author": "Laurie"
-    },
-    {
-        "id": "2",
-        "person": "Olivier",
-        "title": "Rendez-vous RSA",
-        "category": "assistance administrative",
-        "dueDate": "2019-11-21T14:20:44.461Z",
-        "creationDate": "2019-04-04T14:20:44.461Z",
-        "description": "Some description of the task and more details if needed like location, links or whats needed to compelte the task",
-        "author": "Julien",
-        "completedDate": "2019-11-20T14:20:44.461Z",
-        "completedBy": "Tony"
-    },
-    {
-        "id": "3",
-        "person": "Mimi",
-        "title": "Recherche num SS Ameli",
-        "category": "assistance administrative",
-        "dueDate": "2019-12-08T21:43:44.461Z",
-        "creationDate": "2018-12-03T21:43:44.461Z",
-        "description": "Some description of the task and more details if needed like location, links or whats needed to compelte the task",
-        "author": "Laurie"
-    },
-    {
-        "id": "4",
-        "person": "M des Batignolles",
-        "title": "Accompagnement préfécture",
-        "category": "déplacement",
-        "dueDate": "2019-12-22T12:01:44.461Z",
-        "creationDate": "2019-01-12T12:01:44.461Z",
-        "description": "Some description of the task and more details if needed like location, links or whats needed to compelte the task",
-        "author": "Thierry"
-    }
-]
