+++ src/token.js
@@ -1,4 +1,7 @@
+// todo: use Storage
 const getToken = () =>
   "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiZXpmemdmemVmZWYiLCJodHRwczovL2hhc3VyYS5pby9qd3QvY2xhaW1zIjp7IngtaGFzdXJhLWFsbG93ZWQtcm9sZXMiOlsidXNlciJdLCJ4LWhhc3VyYS1kZWZhdWx0LXJvbGUiOiJ1c2VyIiwieC1oYXN1cmEtdXNlci1pZCI6IjhiYWIwZGJkLWU3MzQtNDFmZi05NjYwLWY1MjEyN2U2OGJhMyJ9LCJpYXQiOjE1NzU1MDE4MjEsImV4cCI6MTU3ODA5MzgyMSwic3ViIjoiOGJhYjBkYmQtZTczNC00MWZmLTk2NjAtZjUyMTI3ZTY4YmEzIn0.C7mEg_z-Y6tWWwwzMP74UXk9A4e8_2INDNBIqQRYboNBFVgYdD3xGcHoVMvzhZG7Rk-vX-2YOG2vo4zwtkWOLC-gZ09z3dkTfOs0FsjYBa-y2jO0s7odyBqicNb3MyMsi2gJb4d_rjH5ZNvBFMSogitiL0LleHMeYC0kPC4cNrM4rojjoPbFcLTWkfRaLAmqQTFA7KUyop6eVVGxqt_jWnTyIczAJnNQASH4oIuA-qdrZjU5FTdz_w-Wa53STDIUSGglT6bx7HF7Okywi-o7lD3483WkAxAuKjqtkwE7VZlJdOpQuTQSukOCVgCg1UEr0vUJ_Kam96gh3bVFf0go0g";
 
-export default getToken;
+const setToken = () => true;
+
+export { getToken, setToken };
