--- tslint.json
@@ -1,4 +1,0 @@
-{
-  "extends": ["tslint-react"],
-  "defaultSeverity": "warn"
-}
